simplify entry point to speed up loading time
